Move crate out of pcode repository (#1)
This is a move of the code from
https://github.com/mnemonikr/symbolic-pcode/tree/4a9b6eb20fb0ba7f9e1c8a076288133f9cb811f5/crates/libsla-sys
. The reason for this is that the local sys crate in the workspace will
conflict with any other crate which uses it. Since the Sleigh compiler
uses this crate, and the integration tests use the sleigh compiler, the
tests cannot use the version of the sleigh compiler which uses the sys
crate.

Moving the sys crate into its own repository simplifies the
relationship.