Add support for initializing Sleigh directly from sla file contents (#3)
This change adds a new method to support initializing Sleigh directly
from the sla file contents.

This is a slight modification to the standard Sleigh initialization,
which as of Ghidra 11.1 would parse an XML file to obtain the sla file
path, read the sla file, and then initialize. With this change, the sla
specification no longer needs to exist as a file on disk.