Add initialize_from_raw_sla (#6)
This adds the ability to build the sleigh object using the raw sla
format. The standard sla format includes a 4-byte header: `0x73 0x6c
0x61 0x04`. The first 3 bytes are the characters `sla`, and the last
byte is the format version number (currently 4). The remaining bytes are
the sla data, which are zlib-compresed.

The raw sla format does not include the header and does not compress the
data. One use for this format is for fuzzing. It is possible to build a
fuzzer that compresses the generated input to ensure it decompresses
successfully, but this needless roundtrip will substantially slow down
the fuzzer.