Update build to use C++11 for compatibility with Ghidra (#7)
Ghidra builds its decompiler, of which libsla is a component, using
C++11. So far building with C++14 has worked but there's no reason to be
out-of-sync with Ghidra. This has caused complications when submitting
Ghidra patches since e.g. `std::make_unique` isn't available in C++11,
but the patches work under this library.